requirements.txt
openai-whisper
TTS
sounddevice
pydub
PyAudio
librosa
noisereduce
numpy
python-dotenv